To kill a server_2_kill_not_working

+++ Jenkinsfile
@@ -12,7 +12,13 @@ pipeline {
                     // Find the PID of the Node.js server process
                      sh 'pid=$(ps aux | grep "node server.js" | grep -v grep | awk "{print \$2}")'
                     // Stop the server process using its PID
-                    sh 'kill $pid'
+                    if (pid) {
+                        // Stop the server process using its PID
+                        sh "kill $pid"
+                        echo "Node.js server process (PID: $pid) has been stopped."
+                    } else {
+                        echo "Node.js server process is not running."
+                    }
                 }
             }
         }
